requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
streamlit>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
urllib3>=2.0.0
openai>=1.0.0
websockets>=12.0
streamlit-webrtc>=0.47.0
av>=12.0.0
numpy>=1.24.0
sounddevice>=0.4.6
fastapi=0.128.0
